backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9
openai==1.10.0
anthropic==0.18.0
python-json-logger==2.0.7
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
